+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/assets/domain/asbuilt/repository/AssetAsBuiltRepository.java
@@ -23,5 +23,4 @@ import org.eclipse.tractusx.traceability.assets.domain.base.AssetRepository;
 
 
 public interface AssetAsBuiltRepository extends AssetRepository {
-
 }